Move api subpackage out of the htsget-server
This was accidentally moved one level too deep.
