+++ htsget-server/main.go
@@ -23,7 +23,7 @@ import (
 	"strings"
 
 	"github.com/google/uuid"
-	"github.com/googlegenomics/htsget/htsget-server/api"
+	"github.com/googlegenomics/htsget/api"
 	"github.com/googlegenomics/htsget/internal/analytics"
 )
